moved style restoring code to +page.svekte
